requirements.txt
numpy==1.26.0
Flask==2.2.2
werkzeug==2.2.2
arxiv
opencv-python-headless
Pillow